refactor(types): single shared definition for chat types across main and renderer
ChatMessage/ChatContentBlock/MessageUsage, ToolActivity, ChatTurnSummary and
the permission request/decision shapes were defined twice — once under
electron/ and once hand-mirrored in src/types.ts — with nothing catching
drift between the copies. They now live once in electron/shared/chat-types.ts
(type-only, compiled by both tsconfigs); session-reader/chat-runner re-export
them for existing importers and src/types.ts re-exports them for the renderer.
