+++ src/types.ts
@@ -54,30 +54,19 @@ export interface ClaudeMdHierarchy {
   layers: ClaudeMdLayer[]
 }
 
-export type ChatContentBlock =
-  | { type: 'text'; text: string }
-  | { type: 'thinking'; thinking: string }
-  | { type: 'tool_use'; id: string; name: string; input: Record<string, unknown> }
-  | { type: 'tool_result'; toolUseId: string; content: string; isError: boolean }
-
-export interface ChatMessage {
-  uuid: string
-  role: 'user' | 'assistant'
-  timestamp: string
-  model?: string
-  content: ChatContentBlock[]
-  usage?: MessageUsage
-}
-
-// Token usage di un singolo turno assistant (assente sui turni user). La somma
-// `input + cacheRead + cacheWrite` dell'ultimo turno ≈ l'occupazione corrente
-// del context window, da cui il Mission Control deriva il gauge CONTEXT.
-export interface MessageUsage {
-  inputTokens: number
-  outputTokens: number
-  cacheReadTokens: number
-  cacheWriteTokens: number
-}
+// I tipi della chat (messaggi, stream SDK, permessi) hanno una definizione
+// unica condivisa col main process — vedi electron/shared/chat-types.ts.
+// Ri-esportati qui così il renderer continua a importarli da './types'.
+export type {
+  ChatContentBlock,
+  ChatMessage,
+  MessageUsage,
+  ToolActivity,
+  ChatTurnSummary,
+  PermissionSuggestion,
+  PermissionRequest,
+  PermissionDecision,
+} from '../electron/shared/chat-types'
 
 // A normalized session-lifecycle notification pushed from the main process over
 // `notifications:event`. Mirrors electron/modules/notifications/types.ts (the two
@@ -129,65 +118,6 @@ export interface SessionSummary {
 export interface ExportSaveResult {
   canceled: boolean
   filePath: string | null
-}
-
-/** A `PermissionUpdate` suggestion from the SDK — the rule(s) to persist when the
- *  user picks "Always allow". Shape is opaque to the renderer; it round-trips
- *  back to the SDK verbatim, so we keep it loosely typed. */
-export type PermissionSuggestion = Record<string, unknown>
-
-/** A tool-approval request forwarded from the main process (`canUseTool`). The
- *  renderer renders an Allow / Always / Deny dialog and answers with
- *  `respondPermission(requestId, decision)`. */
-export interface PermissionRequest {
-  requestId: string
-  toolName: string
-  /** Full prompt sentence from the bridge (e.g. "Claude wants to read foo.txt"). */
-  title?: string
-  /** Short noun phrase for the action (e.g. "Read file"). */
-  displayName?: string
-  /** Human-readable subtitle. */
-  description?: string
-  /** The tool input (e.g. `{ command }` for Bash). */
-  input: Record<string, unknown>
-  /** Permission rules to persist on "Always allow". */
-  suggestions?: PermissionSuggestion[]
-  /** Path that triggered the request, when applicable. */
-  blockedPath?: string
-  /** Why the request was triggered. */
-  decisionReason?: string
-  toolUseID: string
-}
-
-/** The renderer's verdict on a `PermissionRequest`, returned to the SDK. */
-export type PermissionDecision =
-  | { kind: 'allow'; input: Record<string, unknown> }
-  | { kind: 'always'; input: Record<string, unknown>; suggestions?: PermissionSuggestion[] }
-  | { kind: 'deny'; message?: string }
-
-/** Live tool indicator for the in-flight turn (`sessions:chatToolActivity`):
- *  emitted when the model starts writing a tool call's input (elapsedSeconds
- *  null) and periodically while the tool runs (elapsedSeconds set, from the
- *  SDK's `tool_progress`). Mirrors `ToolActivity` in `chat-runner.ts`. */
-export interface ToolActivity {
-  toolName: string
-  elapsedSeconds: number | null
-}
-
-/** End-of-turn metadata for the in-app SDK chat, derived from the SDK's `result`
- *  message (NOT from disk) and carried on `sessions:chatDone`. Lets the live chat
- *  show running cost/tokens/model without ever reading the transcript file.
- *  Mirrors `ChatTurnSummary` in `chat-runner.ts`. Cost and token counts are
- *  cumulative for the session (the SDK reports session totals on each result). */
-export interface ChatTurnSummary {
-  totalCostUsd: number
-  inputTokens: number
-  outputTokens: number
-  cacheReadTokens: number
-  cacheWriteTokens: number
-  numTurns: number
-  /** Models used so far (the keys of the SDK's `modelUsage`). */
-  models: string[]
 }
 
 export type ArtifactKind = 'session' | 'subagents' | 'tasks' | 'plan'
